giving this shit second life

+++ application/schemas/payment_schemas.py
@@ -1,14 +1,14 @@
 from uuid import UUID
 
 from pydantic import BaseModel, EmailStr, Field
-
+from decimal import Decimal
 from application.schemas import OrderItemS
 
 
 class CreatePaymentS(BaseModel):
     customer_full_name: str = Field(min_length=3)
     customer_email: EmailStr
-    total_amount: float = Field(ge=1.0)
+    total_amount: Decimal = Field(ge=1.0)
     currency: str = Field(default="RUB", min_length=2)
     description: str
     items: list[OrderItemS] | None = Field(default=None)
@@ -17,9 +17,3 @@ class CreatePaymentS(BaseModel):
 class ReturnPaymentS(BaseModel):
     confirmation_url: str
     payment_id: UUID
-
-
-
-
-
-
