+++ application/schemas/shopping_session_schemas.py
@@ -1,5 +1,6 @@
 from datetime import datetime
 from uuid import UUID
+from decimal import Decimal
 
 from pydantic import BaseModel, Field
 
@@ -12,7 +13,7 @@ class CreateShoppingSessionS(BaseModel):
 class ReturnShoppingSessionS(BaseModel):
     id: UUID
     user_id: int | None
-    total: float
+    total: Decimal
     expiration_time: datetime
